feat/fix: refactoring  multiple files and adding some new cool resources to data_cleaning.py
