+++ test/test_dataCleaning.py
@@ -1,4 +1,4 @@
-from ez_docs.modules.data_cleaning import filter_data
+from ez_docs.modules.data_cleaning import filter_data, filter_format
 
 final_data = [
     {'nome': 'Bruno', 'idade': '18'},
@@ -14,6 +14,12 @@ final_data2 = [
     {'nome': 'Igor', 'idade': '18'},
     {'nome': 'Igor', 'idade': '18'},
 ]
+
+def test_find_delimiter():
+    assert find_delimiter("test/teams1.csv") == ","
+    assert find_delimiter("test/teams2.csv") == ";"
+    assert find_delimiter("test/teams3.csv") == "\\"
+    assert find_delimiter("test/teams4.csv") == "~"
 
 def test_data_cleaning():
     assert filter_data("test/example.csv") == final_data
